Return read-only map as confidence result map
The previously returned SortedMap relied on confidence values for sorting,
so trying to put a Language without confidence value in the result map or
using the SortedMap.comparator() with a Language without confidence value
would cause an exception.

Therefore now switch to a read-only map. While this does not allow using
the convenience method SortedMap.firstKey(), this can easily be worked
around using the key set, e.g. `map.keySet().iterator().next()` (with Kotlin
extension functions this is even easier).