Add hash map based model format
Adds an additional model format where ngrams are stored in primitive hash
maps. Compared to the current binary search implementation this increases
detection speed quite a lot (~24ms to 14ms for short texts; ~51ms to 37ms
for longer texts with single threaded executor). However, with all language
models preloaded it also increases memory usage from ~131 MB to 342 MB.
Therefore this was added as separate mode but the current binary search
based implementation was kept.

This was implemented on top of the binary search based model format to avoid
having to adjust the model file format. While this causes some overhead for
model loading of the hash map based format, it still seems to be reasonably
fast.

Also, the polymorphism (lookups using ...2FloatFunction instead of specific
implementation) does not seem to have a great (or any?) performance impact.
A previous implementation which was using only hash maps had similar
performance.

Resolves #1