Don't build 'JAR with dependencies' by default
This task takes quite some time because it has to include the complete
Kotlin standard lib.

This commit also disables relocation to speed up generation, and to make
it less error-prone (due to not requiring bytecode rewriting).
JAR with dependencies is only intended for CLI usage so no relocation
should be necessary.